pyproject.toml
[project]
name = "level-bot"
version = "0.1.0"
description = "Discord server statistics bot with web dashboard"
readme = "README.md"
license = "MIT"
requires-python = "==3.12.*"
dependencies = [
    "discord.py~=2.6.0",
    "SQLAlchemy~=2.0.0",
    "alembic~=1.18.0",
    "pydantic-settings~=2.0",
    "asyncpg~=0.30.0",
    "greenlet~=3.0.0",
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.34.0",
    "psycopg2-binary~=2.9.0",
    "httpx~=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0",
    "pytest-cov~=4.0",
    "pytest-asyncio~=0.23",
    "mypy~=1.8",
    "ruff~=0.14.0",
    "testcontainers[postgres]~=4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM",
]

[tool.ruff.lint.per-file-ignores]
"src/web/app.py" = ["B008"]
"src/features/**/routes.py" = ["B008"]
"src/**/test_*.py" = ["ARG002"]
"conftest.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
# テストはコロケーション (src/ 内のソースファイル隣に test_*.py を置く)
testpaths = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
concurrency = ["greenlet", "thread"]
# テストはコロケーションされているのでカバレッジから除外する
omit = ["*/migrations/*", "*/test_*.py", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
fail_under = 60
